+++ screens/signup/pages/StartPage.jsx
@@ -1,14 +1,14 @@
-import React, { css, useEffect, useState } from "react";
-import styled from "styled-components/native";
-import { WithLocalSvg } from "react-native-svg";
-import { useNavigation } from "@react-navigation/native";
-import { login } from "@react-native-seoul/kakao-login";
-import NaverLogin from "@react-native-seoul/naver-login";
+import React, { useEffect, useState } from 'react';
+import styled from 'styled-components/native';
+import { WithLocalSvg } from 'react-native-svg';
+import { useNavigation } from '@react-navigation/native';
+import { login } from '@react-native-seoul/kakao-login';
+import NaverLogin from '@react-native-seoul/naver-login';
 
-import FlintLogoSvg from "../../assets/images/flint_logo.svg";
-import KakaoLogoSvg from "../../assets/images/kakao_logo.svg";
-import NaverLogoSvg from "../../assets/images/naver_logo.svg";
-import DetailsInfoPage from "./DetailsInfoPage";
+import FlintLogoSvg from '../../../assets/images/flint_logo.svg';
+import KakaoLogoSvg from '../../../assets/images/kakao_logo.svg';
+import NaverLogoSvg from '../../../assets/images/naver_logo.svg';
+import DetailsInfoPage from './DetailsInfoPage';
 
 const Container = styled.View`
   flex: 1;
@@ -48,28 +48,28 @@ const StartPage = () => {
       const token = await login();
       setToken(token.accessToken);
     } catch (err) {
-      console.error("error: ", err);
+      console.error('error: ', err);
     }
   };
 
   const signInWithNaver = async () => {
     const { failureResponse, successResponse, isSuccess } =
       await NaverLogin.login({
-        appName: "com.flintfront",
-        consumerKey: "to11DuJF425zfdagzPWl",
-        consumerSecret: "MVoE5xJDwh",
-        serviceUrlScheme: "flint",
+        appName: 'com.flintfront',
+        consumerKey: 'to11DuJF425zfdagzPWl',
+        consumerSecret: 'MVoE5xJDwh',
+        serviceUrlScheme: 'flint',
       });
     if (isSuccess) {
       setToken(successResponse.accessToken);
     } else {
-      console.log("error: " + failureResponse);
+      console.log('error: ' + failureResponse);
     }
   };
 
   useEffect(() => {
     navigation.navigate(DetailsInfoPage);
-    console.log("accessToken: " + token);
+    console.log('accessToken: ' + token);
   }, [token]);
 
   return (
@@ -77,22 +77,22 @@ const StartPage = () => {
       <WithLocalSvg width={135} asset={FlintLogoSvg} />
       <BtnWrap>
         <StartBtn
-          backgroundColor={"#fae100"}
+          backgroundColor={'#fae100'}
           onPress={() => {
             signInWithKakao();
           }}
         >
           <WithLocalSvg height={22} asset={KakaoLogoSvg} />
-          <BtnText color={"#381e1f"}>카카오톡으로 시작하기</BtnText>
+          <BtnText color={'#381e1f'}>카카오톡으로 시작하기</BtnText>
         </StartBtn>
         <StartBtn
-          backgroundColor={"#3bac37"}
+          backgroundColor={'#3bac37'}
           onPress={() => {
             signInWithNaver();
           }}
         >
           <WithLocalSvg height={22} asset={NaverLogoSvg} />
-          <BtnText color={"#fff"}>네이버로 시작하기</BtnText>
+          <BtnText color={'#fff'}>네이버로 시작하기</BtnText>
         </StartBtn>
       </BtnWrap>
     </Container>
